Add BuildSceneAttribute
