+++ README.md
@@ -217,10 +217,10 @@ It has the following benefits:
 
 - Can be use same code for running Edit Mode tests, Play Mode tests in Editor, and on Player.
 - Can be specified scenes that are **NOT** in "Scenes in Build".
-- Can be specified path by glob pattern. However, there are restrictions, top level and scene name cannot be omitted.
-- Can be specified path by relative path from the test class file.
+- Can be specified scene path by [glob](https://en.wikipedia.org/wiki/Glob_(programming)) pattern. However, there are restrictions, top level and scene name cannot be omitted.
+- Can be specified scene path by relative path from the test class file.
 
-- This attribute can attached to the test method only.
+This attribute can attached to the test method only.
 It can be used with sync Tests, async Tests, and UnityTest.
 
 Usage:
