+++ Runtime/Attributes/BuildSceneAttribute.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a41ca53cb4a74fb28baa6f0dc03d2cfd
+timeCreated: 1714362832
