+++ Tests/Runtime/Attributes/BuildSceneAttributeTest.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 107c5cb6460248c7912bec5795a0b3c4
+timeCreated: 1714365787
