+++ Tests/Scenes/NotInScenesInBuildForGlob.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 57baa7f9a542e49f39ee705d1cf417e4
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
