+++ Tests/Scenes/NotInScenesInBuildForRelative.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d879e62eb359144bda460d1fa5f0f75d
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
