+++ Tests/Scenes/NotInScenesInBuildForUse.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 04402e0eb57484956bcd85535d57faca
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
